backend/pyproject.toml
[project]
name = "evolution-todo-backend"
version = "0.1.0"
description = "Evolution of Todo — Phase II: FastAPI backend (Hackathon II)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
